requirements.txt
# platform dependent https://pytorch.org/get-started/locally/#start-locally
# torch
# torchvision
# torch-tensorrt

scikit-learn
dill
jsonschema
PyYAML
ruff
mypy
coverage
pytest
parameterized
torchmetrics
numpy
tqdm
uuid
onnxscript
pillow
